SensorHAL: light: correct delay times
Integration time was incorrectly saved as ms instead of us.
And set delay function was doing checks properly us vs us which would end up on a us vs ms check.

Correct that behavior in order to use the proper period which is 1000 times slower and ease cpu usage greatly.